:recycle: refactor: Throw exception when not found Purchase

+++ src/main/java/com/ghostdovahkiin/librapi/purchase/services/DeletePurchaseServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ghostdovahkiin.librapi.purchase.services;
 
+import com.ghostdovahkiin.librapi.exceptions.PurchaseNotFoundException;
 import com.ghostdovahkiin.librapi.purchase.PurchaseRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,6 +12,9 @@ public class DeletePurchaseServiceImpl implements DeletePurchaseService{
 
     @Override
     public void delete(Long id) {
+        if (!purchaseRepository.existsById(id)){
+            throw new PurchaseNotFoundException();
+        }
         purchaseRepository.deleteById(id);
     }
 }
